backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-docx>=1.1
sentence-transformers>=2.7
faiss-cpu>=1.8
rank-bm25>=0.2
httpx>=0.27
numpy>=1.26
